+++ NEWS.md
@@ -1,3 +1,5 @@
+* translated 20200323 picco (#23)
+
 # covid19ita 0.19.0
 
 * spellcheck